LU-18475 build: compatibility updates for kernel 6.12
Linux commit v6.6-rc2-11-gd77008421afd
 groups: Convert group_info.usage to refcount_t
Provide wrappers to inc/dec group_info.usage

Linux v6.12-rc1-3-g5f60d5f6bbc1
 move asm/unaligned.h to linux/unaligned.h
Add a configure test to determine which header to use

Linux v6.11-rc1-51-ga225800f322a
 fs: Convert aops->write_end to take a folio
Linux v6.11-rc1-52-g1da86618bdce
 fs: Convert aops->write_begin to take a folio
Add 'struct folio' for page vs folio signature change.

Linux v6.11-rc4-27-g11068e0b64cb
  fs: remove f_version
f_version is removed, conditionally ignore it.

Linux v6.11-rc6-86-g09022bc196d2
  mm: remove PG_error
PG_error flag and PageError wrappers are removed.

Linux v6.11-rc6-233-g99f86bbda317
  mm: remove PageMlocked
PageMLocked wrappers are removed

Linux v6.11-rc6-225-ge880034cf718
  mm: introduce page_mapcount_is_type()
PAGE_MAPCOUNT_RESERVE is removed and page_mapcount_is_type()
is used instead.

Test-Parameters: trivial
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57125
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/llite/llite_lib.c
@@ -3086,8 +3086,8 @@ void ll_truncate_inode_pages_final(struct inode *inode)
 
 	LASSERTF((inode->i_state & I_FREEING) || inode_is_locked(inode),
 		 DFID ":inode %px state %#lx, lli_flags %#lx\n",
-		 PFID(ll_inode2fid(inode)), inode, inode->i_state,
-		 ll_i2info(inode)->lli_flags);
+		 PFID(ll_inode2fid(inode)), inode,
+		 (unsigned long)inode->i_state, ll_i2info(inode)->lli_flags);
 
 	truncate_inode_pages_final(mapping);
 
@@ -3114,7 +3114,7 @@ void ll_truncate_inode_pages_final(struct inode *inode)
 #endif
 		CWARN("%s: inode="DFID"(%p) nrpages=%lu state %#lx, lli_flags %#lx, see https://jira.whamcloud.com/browse/LU-118\n",
 		      ll_i2sbi(inode)->ll_fsname, PFID(ll_inode2fid(inode)),
-		      inode, nrpages, inode->i_state,
+		      inode, nrpages, (unsigned long)inode->i_state,
 		      ll_i2info(inode)->lli_flags);
 #ifdef HAVE_XARRAY_SUPPORT
 		rcu_read_lock();
